auto fix db files on repack

+++ LpkTool.Library/DbHelpers/ItemHelper.cs
@@ -12,6 +12,11 @@ namespace LpkTool.Library.DbHelpers
         private string TempDir => Path.Combine(Path.GetTempPath(), "lpktool_" + _guid);
         private readonly Lpk _lpk;
 
+        public static ItemHelper FromFile(string lpkPath)
+        {
+            return new ItemHelper(Lpk.FromFile(lpkPath));
+        }
+
         /// <summary>
         /// 
         /// </summary>
@@ -59,7 +64,8 @@ namespace LpkTool.Library.DbHelpers
             Microsoft.Data.Sqlite.SqliteConnection.ClearAllPools();
             GC.Collect();
             GC.WaitForPendingFinalizers();
-            Directory.Delete(TempDir, true);
+            if (Directory.Exists(TempDir))
+                Directory.Delete(TempDir, true);
             GC.SuppressFinalize(this);
         }
 
@@ -85,6 +91,177 @@ namespace LpkTool.Library.DbHelpers
                 connection.Dispose();
             }
             return result;
+        }
+
+        private List<string> KeyWhitelist = new List<string> {
+            "PrimaryKey",
+            "Name",
+            "Comment",
+            "Icon",
+            "IconIndex",
+            "IconSound",
+            "DropModel",
+            "Model",
+            "Type",
+            "Grade"
+        };
+
+        public List<Dictionary<string, object>> GetOutfitsForClass(string classAttribute, int grade = -1)
+        {
+            var allItems = GetAllItems(true);
+            var outfits = allItems.Where(x => x.Value[classAttribute].ToString() == "1" && x.Value["Type"].ToString() == "9").Select(x => x.Value).OrderBy(x => x["Icon"]).ToList();
+            if (grade > -1)
+            {
+                outfits = outfits.Where(x => x["Grade"].ToString() == grade.ToString()).ToList();
+            }
+            foreach (var outfit in outfits)
+            {
+                foreach (var item in outfit.Where(x => !(KeyWhitelist.Contains(x.Key) || (x.Key.StartsWith("For") && x.Value.ToString() == "1"))))
+                {
+                    outfit.Remove(item.Key);
+                }
+            }
+            return outfits;
+        }
+
+        public void RepackToFile(string outPath)
+        {
+            var itemFile = _lpk.GetFileByName("EFTable_Item.db");
+            itemFile.ReplaceData(this.GetFileData("EFTable_Item.db"));
+            _lpk.RepackToFile(outPath);
+        }
+
+        public void ModelSwap(int targetPrimaryKey, int sourcePrimaryKey)
+        {
+            var tempFile = Path.Combine(TempDir, "EFTable_Item.db");
+            if (!File.Exists(tempFile))
+            {
+                File.WriteAllBytes(tempFile, _lpk.GetFileByName("EFTable_Item.db")!.GetData());
+            }
+
+            using (var connection = new SqliteConnection($"Data Source={tempFile}"))
+            {
+                connection.Open();
+                var command = connection.CreateCommand();
+                command.CommandText = $"SELECT Model From Item WHERE PrimaryKey = {sourcePrimaryKey}";
+                string model = "";
+                using (var reader = command.ExecuteReader())
+                {
+                    if (reader.Read())
+                    {
+                        model = reader.GetString(0);
+                    }
+                    reader.Close();
+                    reader.DisposeAsync().GetAwaiter().GetResult();
+                }
+
+                if (!string.IsNullOrEmpty(model))
+                {
+                    var (query, newValues) = GetModelSwapQuery(targetPrimaryKey, model);
+                    var updateCommand = BuildUpdateCommand("Item", query, newValues);
+                    var command2 = connection.CreateCommand();
+                    command2.CommandText = updateCommand;
+                    command2.ExecuteNonQuery();
+                }
+
+                connection.Close();
+                connection.Dispose();
+            }
+
+        }
+
+        public void ModelSwap(int targetPrimaryKey, string model)
+        {
+            var (query, newValues) = GetModelSwapQuery(targetPrimaryKey, model);
+            UpdateValuesForItemTable(query, newValues);
+        }
+
+        public static (List<Dictionary<string, object>> query, Dictionary<string, object> newValues) GetModelSwapQuery(int primaryKey, string newModel)
+        {
+            var query = new List<Dictionary<string, object>>();
+            var newValues = new Dictionary<string, object>();
+            query.Add(new Dictionary<string, object>() { { "PrimaryKey", primaryKey } });
+            newValues.Add("Model", newModel);
+            return (query, newValues);
+        }
+
+        public void UpdateValuesForItemTable(List<Dictionary<string, object>> query, Dictionary<string, object> newValues)
+        {
+            UpdateValues("EFTable_Item.db", "Item", query, newValues);
+        }
+
+        public void UpdateValues(string fileName, string tableName, List<Dictionary<string, object>> query, Dictionary<string, object> newValues)
+        {
+            var tempFile = Path.Combine(TempDir, fileName);
+            if (!File.Exists(tempFile))
+            {
+                File.WriteAllBytes(tempFile, _lpk.GetFileByName(fileName)!.GetData());
+            }
+
+            using (var connection = new SqliteConnection($"Data Source={tempFile}"))
+            {
+
+                var updateCommand = BuildUpdateCommand(tableName, query, newValues);
+                connection.Open();
+                var command = connection.CreateCommand();
+                command.CommandText = updateCommand;
+                command.ExecuteNonQuery();
+                connection.Close();
+                connection.Dispose();
+            }
+
+        }
+
+        public byte[] GetFileData(string fileName)
+        {
+            Microsoft.Data.Sqlite.SqliteConnection.ClearAllPools();
+            GC.Collect();
+            var tempFile = Path.Combine(TempDir, fileName);
+            if (!File.Exists(tempFile))
+            {
+                File.WriteAllBytes(tempFile, _lpk.GetFileByName(fileName)!.GetData());
+            }
+            return File.ReadAllBytes(tempFile);
+        }
+
+        private string BuildUpdateCommand(string tableName, List<Dictionary<string, object>> query, Dictionary<string, object> newValues)
+        {
+            var snippet = "{0} = {1}";
+            var command = $"UPDATE {tableName}";
+            var setLine = "SET ";
+            for (int i = 0; i < newValues.Count; i++)
+            {
+                var item = newValues.ElementAt(i);
+                if (i != 0)
+                {
+                    setLine += ", ";
+                }
+                setLine += string.Format(snippet, item.Key, (item.Value is string ? $"'{item.Value}'" : item.Value));
+            }
+            setLine = setLine.TrimEnd(new char[] { ',', ' ' });
+            var whereLine = "WHERE ";
+            for (var i = 0; i < query.Count; i++)
+            {
+                var queryCombo = query[i];
+                if (i != 0)
+                {
+                    whereLine += " OR ";
+                }
+                whereLine += "(";
+                for (int j = 0; j < queryCombo.Count; j++)
+                {
+                    var item = queryCombo.ElementAt(j);
+                    if (j != 0)
+                    {
+                        whereLine += " AND ";
+                    }
+                    whereLine += string.Format(snippet, item.Key, (item.Value is string ? $"'{item.Value}'" : item.Value));
+                }
+                whereLine += ")";
+            }
+            whereLine = whereLine.TrimEnd(new char[] { ',', ' ' });
+            command = $"{command} {setLine} {whereLine}";
+            return command;
         }
 
         private void AddCommandEvaluationToResult(SqliteCommand command, ref Dictionary<int, Dictionary<string, object>> result, bool resolvedNames)
@@ -152,11 +329,6 @@ namespace LpkTool.Library.DbHelpers
             var nameKeys = GetMessageKeysContainingName(name);
             var items = GetItemsByNameKeys(nameKeys);
             return items;
-        }
-
-        public void SetModel(string model)
-        {
-
         }
 
         private readonly int _limit = 499;
